updated package imports renamed RealDataSearchResultsProvider to DataSearchResultsProvider removed Thread.sleep from controller
