+++ explorer/core/src/main/java/net/catenax/explorer/core/webui/service/DataSearchResultsProvider.java
@@ -3,12 +3,12 @@ package net.catenax.explorer.core.webui.service;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import net.catenax.explorer.core.ExplorerService;
-import net.catenax.explorer.core.submodel.twinregistry.ShellDescriptorResponse;
+import net.catenax.explorer.core.submodel.ShellDescriptorResponse;
 
 import java.util.List;
 
 @RequiredArgsConstructor
-public class RealDataSearchResultsProvider implements SearchResultsProvider {
+public class DataSearchResultsProvider implements SearchResultsProvider {
     private final ExplorerService explorerService;
 
     @Override
