+++ explorer/core/src/main/java/net/catenax/explorer/core/webui/service/SearchResultsProvider.java
@@ -1,6 +1,6 @@
 package net.catenax.explorer.core.webui.service;
 
-import net.catenax.explorer.core.submodel.twinregistry.ShellDescriptorResponse;
+import net.catenax.explorer.core.submodel.ShellDescriptorResponse;
 
 import java.util.List;
